+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 
   * Delay option for request mocks
   * Auto option for request mocks
-  * Trigger method for request mocks 
+  * Trigger method for request mocks
+  * UserAgent option in createBrowser settings
 
 0.8.0 / 2018-06-15
 ==================